backend/requirements.txt
Flask==2.2.3 # was 2.3.2
# Flask-SQLAlchemy==3.0.2
# SQLAlchemy==2.0.16
# pymysql==1.0.2
pandas==1.5.3
numpy==1.23.5
mysql-connector-python==8.0.33 # was no version at first
openpyxl
Werkzeug==2.3.6

